test(ssr): share node runtime smoke coverage
